+++ openfasoc/generators/glayout/glayout/flow/blocks/diffpair_cmirror_bias/diff_pair_cmirrorbias.py
@@ -4,11 +4,11 @@ from gdsfactory.component_reference import ComponentReference
 from gdsfactory.components.rectangle import rectangle
 from glayout.flow.pdk.mappedpdk import MappedPDK
 from typing import Optional, Union
-from glayout.flow.components.blocks.diff_pair import diff_pair
-from glayout.flow.components.primitives.fet import nmos, pmos, multiplier
-from glayout.flow.components.primitives.guardring import tapring
-from glayout.flow.components.primitives.mimcap import mimcap_array, mimcap
-from glayout.flow.components.primitives.via_gen import via_stack, via_array
+from glayout.flow.blocks.diff_pair import diff_pair
+from glayout.flow.primitives.fet import nmos, pmos, multiplier
+from glayout.flow.primitives.guardring import tapring
+from glayout.flow.primitives.mimcap import mimcap_array, mimcap
+from glayout.flow.primitives.via_gen import via_stack, via_array
 from glayout.flow.routing.L_route import L_route
 from glayout.flow.routing.c_route import c_route
 from gdsfactory.routing.route_quad import route_quad
@@ -36,7 +36,7 @@ from glayout.flow.pdk.util.snap_to_grid import component_snap_to_grid
 from pydantic import validate_arguments
 from glayout.flow.placement.two_transistor_interdigitized import two_nfet_interdigitized
 from glayout.flow.spice import Netlist
-from glayout.flow.components.blocks.current_mirror import cmirror_netlist
+from glayout.flow.blocks.current_mirror import cmirror_netlist
 
 def diff_pair_ibias_netlist(center_diffpair: Component, current_mirror: Component, antenna_diode: Optional[Component] = None) -> Netlist:
     netlist = Netlist(
