+++ openfasoc/generators/glayout/glayout/flow/blocks/opamp/opamp.py
@@ -4,13 +4,13 @@ from gdsfactory.component_reference import ComponentReference
 from gdsfactory.components.rectangle import rectangle
 from glayout.flow.pdk.mappedpdk import MappedPDK
 from typing import Optional, Union
-from glayout.flow.components.primitives.fet import nmos, pmos, multiplier
-from glayout.flow.components.blocks.diff_pair import diff_pair
-from glayout.flow.components.primitives.guardring import tapring
-from glayout.flow.components.primitives.mimcap import mimcap_array, mimcap
+from glayout.flow.primitives.fet import nmos, pmos, multiplier
+from glayout.flow.blocks.diff_pair import diff_pair
+from glayout.flow.primitives.guardring import tapring
+from glayout.flow.primitives.mimcap import mimcap_array, mimcap
 from glayout.flow.routing.L_route import L_route
 from glayout.flow.routing.c_route import c_route
-from glayout.flow.components.primitives.via_gen import via_stack, via_array
+from glayout.flow.primitives.via_gen import via_stack, via_array
 from gdsfactory.routing.route_quad import route_quad
 from glayout.flow.pdk.util.comp_utils import evaluate_bbox, prec_ref_center, movex, movey, to_decimal, to_float, move, align_comp_to_port, get_padding_points_cc
 from glayout.flow.pdk.util.port_utils import rename_ports_by_orientation, rename_ports_by_list, add_ports_perimeter, print_ports, set_port_orientation, rename_component_ports
@@ -20,8 +20,8 @@ from pydantic import validate_arguments
 from glayout.flow.placement.two_transistor_interdigitized import two_nfet_interdigitized
 from glayout.flow.spice import Netlist
 
-from glayout.flow.components.blocks.opamp import opamp_twostage
-from glayout.flow.components.blocks.current_mirror import cmirror_netlist
+from glayout.flow.blocks.opamp import opamp_twostage
+from glayout.flow.blocks.current_mirror import cmirror_netlist
 
 def opamp_output_stage_netlist(pdk: MappedPDK, output_amp_fet_ref: ComponentReference, biasParams: list) -> Netlist:
     bias_netlist = cmirror_netlist(pdk, biasParams[0], biasParams[1], biasParams[2])
