+++ openfasoc/generators/glayout/glayout/flow/primitives/guardring.py
@@ -3,7 +3,7 @@ from gdsfactory.cell import cell
 from gdsfactory.component import Component
 from gdsfactory.components.rectangle import rectangle
 from gdsfactory.components.rectangular_ring import rectangular_ring
-from glayout.flow.components.primitives.via_gen import via_array, via_stack
+from glayout.flow.primitives.via_gen import via_array, via_stack
 from typing import Optional
 from glayout.flow.pdk.util.comp_utils import to_decimal, to_float, evaluate_bbox
 from glayout.flow.pdk.util.port_utils import print_ports
