+++ openfasoc/generators/glayout/glayout/flow/routing/L_route.py
@@ -3,7 +3,7 @@ from gdsfactory.component import Component
 from gdsfactory.port import Port
 from glayout.flow.pdk.mappedpdk import MappedPDK
 from typing import Optional, Union
-from glayout.flow.components.primitives.via_gen import via_stack, via_array
+from glayout.flow.primitives.via_gen import via_stack, via_array
 from glayout.flow.pdk.util.comp_utils import evaluate_bbox, align_comp_to_port, to_decimal, to_float, prec_ref_center, get_primitive_rectangle
 from glayout.flow.pdk.util.port_utils import rename_ports_by_orientation, rename_ports_by_list, print_ports, assert_port_manhattan, assert_ports_perpindicular
 from decimal import Decimal
