+++ openfasoc/generators/glayout/glayout/flow/routing/straight_route.py
@@ -3,7 +3,7 @@ from gdsfactory.component import Component
 from gdsfactory.port import Port
 from glayout.flow.pdk.mappedpdk import MappedPDK
 from typing import Optional
-from glayout.flow.components.primitives.via_gen import via_stack, via_array
+from glayout.flow.primitives.via_gen import via_stack, via_array
 from gdsfactory.components.rectangle import rectangle
 from glayout.flow.pdk.util.comp_utils import evaluate_bbox, align_comp_to_port
 from glayout.flow.pdk.util.port_utils import assert_port_manhattan, set_port_orientation, add_ports_perimeter
